Create a new HUD Controller script
Remove the old HUD functionality and move it to the new controller
