+++ Swamp Sneak/Assets/Scripts/Classes/HUDController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: cd393cee7ee500a478279fae1a63cde5
-timeCreated: 1521392043
+guid: d47e164f3c3b5644dbc69824c8e2abbf
+timeCreated: 1523248678
 licenseType: Free
 MonoImporter:
   externalObjects: {}
